+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - There now is a new menu entry in the _Todo menu_ to open the **Nextcloud Deck dialog**
   for editing cards from the stack (for [#3357](https://github.com/pbek/QOwnNotes/issues/3357))
+  - It's now also possible to search for cards in the dialog
 
 ## 25.9.0